+++ projects/rocprofiler-compute/tests/test_profile_general.py
@@ -2370,3 +2370,105 @@ class TestSetsIntegration:
         # workload dir should not exist
         assert not Path(workload_dir).exists()
         test_utils.clean_output_dir(config["cleanup"], workload_dir)
+
+
+@pytest.mark.iteration_multiplexing
+def test_profiler_options(binary_handler_profile_rocprof_compute):
+    options = ["--no-native-tool", "--iteration-multiplexing"]
+    workload_dir = test_utils.get_output_dir()
+    code = binary_handler_profile_rocprof_compute(
+        config, workload_dir, options, check_success=False, roof=False
+    )
+    assert code == 1
+
+
+@pytest.mark.iteration_multiplexing
+def test_iteration_multiplexing(binary_handler_profile_rocprof_compute):
+    options = ["--iteration-multiplexing"]
+    workload_dir = test_utils.get_output_dir()
+    _ = binary_handler_profile_rocprof_compute(
+        config, workload_dir, options, check_success=True, roof=False
+    )
+
+    file_dict = test_utils.check_csv_files(workload_dir, num_devices, num_kernels)
+    if soc == "MI100":
+        assert sorted(list(file_dict.keys())) == CSVS
+    elif soc == "MI200":
+        assert sorted(list(file_dict.keys())) == CSVS
+    elif "MI300" in soc:
+        assert sorted(list(file_dict.keys())) == CSVS
+    elif "MI350" in soc:
+        assert sorted(list(file_dict.keys())) == CSVS
+    else:
+        print(f"Testing isn't supported yet for {soc}")
+        assert 0
+
+    validate(
+        inspect.stack()[0][3],
+        workload_dir,
+        file_dict,
+    )
+
+    test_utils.clean_output_dir(config["cleanup"], workload_dir)
+
+
+@pytest.mark.iteration_multiplexing
+def test_iteration_multiplexing_kernel(binary_handler_profile_rocprof_compute):
+    options = ["--iteration-multiplexing", "kernel"]
+    workload_dir = test_utils.get_output_dir()
+    _ = binary_handler_profile_rocprof_compute(
+        config, workload_dir, options, check_success=True, roof=False
+    )
+
+    file_dict = test_utils.check_csv_files(workload_dir, num_devices, num_kernels)
+    if soc == "MI100":
+        assert sorted(list(file_dict.keys())) == CSVS
+    elif soc == "MI200":
+        assert sorted(list(file_dict.keys())) == CSVS
+    elif "MI300" in soc:
+        assert sorted(list(file_dict.keys())) == CSVS
+    elif "MI350" in soc:
+        assert sorted(list(file_dict.keys())) == CSVS
+    else:
+        print(f"Testing isn't supported yet for {soc}")
+        assert 0
+
+    validate(
+        inspect.stack()[0][3],
+        workload_dir,
+        file_dict,
+    )
+
+    test_utils.clean_output_dir(config["cleanup"], workload_dir)
+
+
+@pytest.mark.iteration_multiplexing
+def test_iteration_multiplexing_kernel_launch_params(
+    binary_handler_profile_rocprof_compute,
+):
+    options = ["--iteration-multiplexing", "kernel_launch_params"]
+    workload_dir = test_utils.get_output_dir()
+    _ = binary_handler_profile_rocprof_compute(
+        config, workload_dir, options, check_success=True, roof=False
+    )
+
+    file_dict = test_utils.check_csv_files(workload_dir, num_devices, num_kernels)
+    if soc == "MI100":
+        assert sorted(list(file_dict.keys())) == CSVS
+    elif soc == "MI200":
+        assert sorted(list(file_dict.keys())) == CSVS
+    elif "MI300" in soc:
+        assert sorted(list(file_dict.keys())) == CSVS
+    elif "MI350" in soc:
+        assert sorted(list(file_dict.keys())) == CSVS
+    else:
+        print(f"Testing isn't supported yet for {soc}")
+        assert 0
+
+    validate(
+        inspect.stack()[0][3],
+        workload_dir,
+        file_dict,
+    )
+
+    test_utils.clean_output_dir(config["cleanup"], workload_dir)