+++ projects/rocprofiler-compute/CHANGELOG.md
@@ -13,6 +13,12 @@ Full documentation for ROCm Compute Profiler is available at [https://rocm.docs.
     * ``--no-native-tool`` option is provided, forcing usage of the default profiler.
     * When performing a dynamic attach to a process for profiling.
 
+* Iteration multiplexing to collect counters in single application run:
+  * Is incompatible with --no-native-tool
+  * Two options:
+    * kernel: Counters are collected in a round robin fashion for unique kernels.
+    * kernel_launch_params: Counters are collected in a round robin fashion for unique kernels having the exact same launch parameters.
+
 ### Changed
 
 * Default output format for the underlying ROCprofiler-SDK tool has been changed from ``csv`` to ``rocpd``.